pyproject.toml
[project]
name = "pavemark_lite"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["torch","torchvision","numpy","pillow","hydra-core"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
